requirements.txt
langchain
langchain-openai
python-dotenv~=1.2.1

pandas~=3.0.1